WTA #79: updated view model to create extract and keep track of creation progress

+++ home/src/main/java/com/jacob/wakatimeapp/home/domain/models/ExtractCreationProgress.kt
@@ -6,7 +6,7 @@ internal data class ExtractCreationProgress(
     val hasFailed: Boolean,
     val isProcessing: Boolean,
     val isStuck: Boolean,
-    val percentComplete: Double,
+    val percentComplete: Float,
     val id: String,
     val expires: String,
     val status: String,
